Add if then statement (#25)
* feat: add parser to support if-then statements

Refactor the parser to support parsing if-then statements. This includes
the addition of `IfThenStatement` class and modifications in the `parseStatement`
method to handle if-then statements. Update `evaluate` method in `Runtime` to
use `parseStatement`. Add a test case in `ParserTest` for parsing if-then
statements.

* feat: Add EqualityExpression class and test method for equality expressions

Added EqualityExpression class with left, operator, and right properties
Implemented parseEqualityExpression method in Parser to handle equality expressions
Updated ParserTest with equalityExpressionTest to test equality expressions
Made changes to ifThenStatementTest in ParserTest to include equality expressions and DeclareVariableStatement class in assertTreeOf function

* feat: Add evaluation for EqualityExpression in Runtime

Added functionality to evaluate EqualityExpression in the Runtime class
to determine if two values are equal by comparing their string representations.

* feat: Add support for If-Then statements

Added the ability to evaluate If-Then statements in the Runtime class. 
If the condition is true, the statement is evaluated, otherwise an empty string is returned.