feat: Add support while statements
This commit adds support for parsing and creating Abstract Syntax Trees
for while statements in the parser. While statements consist of a
conditional expression and a body enclosed in curly braces. This change
allows the parser to handle while statements during the parsing process,
enabling the generation of ASTs for while loops.