Refactor lexer parser fixes (#28)
* refactor: Lexer and Parser rewrite base on Java8.g4

This commit renames the method for evaluating local variable declarations from
'evaluateDeclareVariableStatement' to 'evaluateDeclareVariableStatement' for
clarity and consistency.

* feat: Add BlockStatement parsing and evaluation

Add parsing and evaluation logic for BlockStatement in the Parser and Runtime
classes. This allows for multiple statements to be executed in a block within
curly braces. Update the necessary methods for parsing and evaluating
BlockStatements. Fixes issues related to block statements in the codebase.

* feat: Add support while statements

This commit adds support for parsing and creating Abstract Syntax Trees
for while statements in the parser. While statements consist of a
conditional expression and a body enclosed in curly braces. This change
allows the parser to handle while statements during the parsing process,
enabling the generation of ASTs for while loops.