Method invocation expression implementation  (#30)
* feat: Add method invocation expression support

Added support for method invocation expressions in the Parser. Now the Parser class can parse method invocation expressions with identifiers and arguments.

* feat: Refactor Environment class and add new methods

- Added getIdentifier and getArguments methods to MethodInvocationExpression
- Changed assignment method in Environment to accept BindObject
- Refactored call method in Environment to reuse logic in private methods
- Added isDeclare method to Environment to check if a field is declared
- Added access method to Environment for accessing fields
- Updated Environment constructor to accept a parent environment
- Changed declare method in Environment to accept BindObject
- Updated evaluate method in Runtime to return BindObject
- Refactored evaluate methods in Runtime to return BindObject
- Updated evaluateIfThenStatement method in Runtime to return BindObject

* refactor: Improve BindObject constructor and arithmetic methods

Update the BindObject constructor to include the identifier and modify
arithmetic methods to use the identifier in new BindObject instances for
better tracking of the source of the operation.

* feat: Add pre and post increment and decrement expressions evaluation

Added the evaluation of pre and post increment and decrement expressions to
support operations like `++` and `--`. Updated the `Runtime` class with the
implementation of these evaluations to correctly handle increment and
decrement operations on identifiers.

* feat: Add method declaration statement parsing support

Added support for parsing method declaration statements in the Parser class.
Included method declaration statement parsing functionality in the parseStatement
method to handle method declarations followed by the block statement as the
method body. Updated the parseMethodDeclarationStatement method to correctly
parse method declarations with return type, method name, parameters, and body
statements.

* refactor: Update method declaration parsing in Parser

Change how method declarations are parsed in the Parser class to address
issues related to void return types. Modified the logic to handle void return
types more accurately when checking for return statements and missing return
values. Updated the parsing of BlockStatements and ReturnStatements to align
with the void return type requirements. Added checks for method declarations
with void return types to prevent returning a value and updated the logic to
return the correct type of AST. The error method in Parser was also refactored
to handle both regular and panic errors.

* fix: Update Runtime class to return correct evaluation result

Refactored the evaluate method in the Runtime class to return the message
of the evaluation instead of just the value. Updated method return types
to ShellResult and ResultObject for more accurate functionality. Made
adjustments to evaluation logic in different Statement types for
correct output. Updated corresponding test cases in RuntimeTest.

Fixes: #123

* feat: Update Parser & MethodDeclarationStatement

Add import for Parameter class to Parser.
Change List<Token> to List<Parameter> in Parser method parameters.
Modify Parser to create new Parameter instances and add to method parameters.
Update MethodInvocationExpression parsing for comma-separated arguments.
Add import for Parameter class to MethodDeclarationStatement.
Change List<Token> to List<Parameter> in MethodDeclarationStatement method parameters.
Adjust MethodDeclarationStatement to properly add Parameter elements.
Add getters for returnType, methodName, parameters, and body in MethodDeclarationStatement class.